This is the functioning UI Project with events
The save does not work because the JSON objects don't work the way I want them to.  However, I don't think the requirements included that we should save the new readings, just upload and change status.

Let me know if you have trouble with this.  This is in ZIP so that I make sure all of my stuff gets to you.  Let me know if it doesn't work.  I expect one of you will turn this in.

Also let me know if you need something else from me.   I'll try to get it done, but I'm on vacation in europe so timezones are an issue.